Check infile arguments for BitcodeWrapperer.
It would be better to use a const reference for infile, but unfortunately, we
must mutate it during the lifetime of BitcodeWrapperer - Seek() is the
primary issue here. Since we don't want to use a non-const reference, we
instead insert appropriate NULL checks into the various functions where
infile is used.

Change-Id: Ic408c9dba684b57237cc8e29fff61c103161f7b3